requirements.txt
Django==2.2.2
djangorestframework==3.11.0
djangorestframework-xml==1.4.0
gunicorn==20.0.4
drf-tracking==1.5.0